feat(lab3): add unit tests and enable coverage reporting
- add coverlet.msbuild to NetSdrClientAppTests
- enable Tests with coverage step in sonarcloud.yml
- add NetSdrMessageHelper tests: TranslateMessage_UnknownItemCode_ReturnsFalse, GetSamples_OddByteCount_IgnoresRemainder, GetControlItemMessage_PreservesParameterBytes, TranslateMessage_ControlItemRange_ReturnsCorrectType
- add NetSdrClient tests: ChangeFrequencyTest, StartIQThenStopIQTest
- fix Enum.IsDefined cast from ushort to int in NetSdrMessageHelper.TranslateMessage

+++ NetSdrClientApp/Messages/NetSdrMessageHelper.cs
@@ -83,7 +83,7 @@ namespace NetSdrClientApp.Messages
                 msgEnumarable = msgEnumarable.Skip(_msgControlItemLength);
                 msgLength -= _msgControlItemLength;
 
-                if (Enum.IsDefined(typeof(ControlItemCodes), value))
+                if (Enum.IsDefined(typeof(ControlItemCodes), (int)value))
                 {
                     itemCode = (ControlItemCodes)value;
                 }
